adjusted some broken scripts to allow user dev environment
The delete-db script had the same code written twice, which served no purpose and would crash gulp. init-db now references the directory that it is run from, so regardless of where the gulp is executed from, it will initialize the database correctly. Quick change to a typo in the BOLO status attribute. RECOVER is now RECOVERED and CANCEL is now CANCELLED. init-db now also initializes a default agency with a name, was NULL before, which caused errors when the application attempted to read the agencies. Next step is to play around with the app, see what might be broken, then sprint plan with the stakeholders